Apple Stock Analysis Chatbot/requirements.txt
pandas
yfinance
matplotlib
gradio
google-generativeai
re
word2number